// src/types/types.orbis.d.ts
interface IOrbisGroup {
  channels: Pick<IOrbisChannel, 'content' | 'stream_id'>[]
  content: {
    name?: string
    pfp?: string
    description?: string
  }
  count_members: number
  creator: string
  last_activity_timestamp: number
  stream_id: string
}

interface IOrbisChannel {
  archived?: boolean
  content: {
    group_id: string
    name: string
    description: string
    type: 'chat' | 'feed'
    encryptionRules?: {
      type: string
      chain: string
      contractType: 'ERC20' | 'ERC721' | 'ERC1155'
      contractAddress: string
      minTokenBalance: string
      tokenId?: string
    }
    data?: object
  }
  created_at: string
  creator: string
  group_id: string
  stream_id: string
  timestamp: number
  type: 'chat' | 'feed'
}

export interface IOrbisProfile {
  address: string
  count_followers: number
  count_following: number
  details: {
    a_r?: number
    did: string
    metadata?: {
      address?: string
      chain?: string
      ensName?: string
    }
    nonces?: any
    profile?: {
      cover?: string
      data?: object
      description?: string
      pfp?: string
      pfpIsNft?: {
        chain: string
        contract: string
        timestamp: string
        tokenId: string
      }
      username?: string
    }
    twitter_details?: any
  }
  did: string
  last_activity_timestamp: number
  username: string
}

export interface IOrbisEncryptionRules {
  type: 'token-gated' | 'custom'
  chain: string
  contractType: 'ERC20' | 'ERC721' | 'ERC1155'
  contractAddress: string
  minTokenBalance: string
  tokenId: string
  accessControlConditions?: object
}

export interface IOrbisEncryptedBody {
  accessControlConditions: string
  encryptedString: string
  encryptedSymmetricKey: string
}

export interface IOrbisPostMention {
  did: string
  username: string
}

export interface IOrbisPostContent {
  body: string
  title?: string
  context?: string
  master?: string
  mentions?: IOrbisPostMention[]
  reply_to?: string
  type?: string
  tags?: {
    slug: string
    title: string
  }[]
  data?: object
  encryptionRules?: IOrbisEncryptionRules | null
  encryptedMessage?: object | null
  encryptedBody?: IOrbisEncryptedBody | null
}

export interface IOrbisPost {
  content: IOrbisPostContent
  context?: string
  context_details?: {
    channel_details?: IOrbisChannel['content']
    channel_id?: string
    group_details?: IOrbisGroup['content']
    group_id?: string
  }
  count_commits?: number
  count_downvotes?: number
  count_haha?: number
  count_likes?: number
  count_replies?: number
  creator: string
  creator_details?: IOrbisProfile['details']
  group_id?: string | null
  indexing_metadata?: {
    language?: string
    urlMetadata?: Record<string, string>
  }
  master?: string | null
  reply_to?: string | null
  reply_to_creator_details?: Pick<
    IOrbisProfile['details'],
    'did' | 'metadata' | 'profile'
  >
  reply_to_details?: IOrbisPostContent
  stream_id: string
  timestamp: number
  type?: string
}